api_server: don't wait for machined registration
Not all pods will register, and this slows us down a lot for those pods
that won't.
